refactor: 외부 스크립트 컴포넌트로 변경
