+++ components/scripts/KakaoDeveloper.tsx
@@ -1,0 +1,10 @@
+const KakaoDeveloper = () => {
+    return (
+        <script
+            defer
+            src="https://developers.kakao.com/sdk/js/kakao.min.js"
+        ></script>
+    );
+};
+
+export default KakaoDeveloper;
