+++ components/scripts/index.tsx
@@ -1,0 +1,2 @@
+export { default as GoogleAnalytics } from './GoogleAnalytics';
+export { default as KakaoDeveloper } from './KakaoDeveloper';
