+++ pages/_document.tsx
@@ -1,13 +1,11 @@
 import { Head, Html, Main, NextScript } from 'next/document';
+import { KakaoDeveloper } from '@components/scripts';
 
 export default function Document() {
     return (
         <Html>
             <Head>
-                <script
-                    defer
-                    src="https://developers.kakao.com/sdk/js/kakao.min.js"
-                ></script>
+                <KakaoDeveloper />
             </Head>
             <body>
                 <Main />
