+++ shared/utils/gtag.ts
@@ -1,17 +1,24 @@
+import { URL } from 'url';
+
+interface GtagEventType {
+    action: string;
+    category: string;
+    label: string;
+    value: number;
+}
+
 export const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GA_ID as string;
 
-// https://developers.google.com/analytics/devguides/collection/gtagjs/pages
-export const pageview = (url: string) => {
-  window.gtag('config', GA_TRACKING_ID, {
-    page_path: url,
-  })
-}
+export const pageview = (url: URL) => {
+    window.gtag('config', GA_TRACKING_ID, {
+        page_path: url,
+    });
+};
 
-// https://developers.google.com/analytics/devguides/collection/gtagjs/events
-export const event = ({ action, category, label, value }: any) => {
-  window.gtag('event', action, {
-    event_category: category,
-    event_label: label,
-    value: value,
-  })
-}
+export const event = ({ action, category, label, value }: GtagEventType) => {
+    window.gtag('event', action, {
+        event_category: category,
+        event_label: label,
+        value: value,
+    });
+};
